Add feature
This edit will show what the bot is working on. It will print into the console when it logs into reddit, starts getting comments, when it gives someone a tip, when it receives a tip (also whether it has found the tip amount or not). This is a sample output:

17:56:53: Logging into reddit...
17:56:57: Getting comments...
17:57:07: Tip has been sent!
18:57:09: Tip Received - Amount Verified

I have been running this feature for a few hours without any errors.